requirements.txt
opencv-python>=4.10.0.84
numpy>=1.24.0
torch>=2.0.0
transformers>=4.30.0
Pillow>=9.0.0

